fix: Update build settings
