+++ src/components/atoms/IconSvg/SvgInfo.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react'
 import { SVGProps } from 'react'
 
 const SvgInfo = (props: SVGProps<SVGSVGElement>) => (
