+++ src/components/const/common.ts
@@ -8,3 +8,13 @@ export const ROTATE = {
 } as const
 
 export type Rotate = keyof typeof ROTATE
+
+export const ICON_KIND = {
+  INFO: 'INFO',
+  OPEN_IN_NEW: 'OPEN_IN_NEW',
+  NAVIGATE_NEXT: 'NAVIGATE_NEXT',
+  MENU: 'MENU',
+  MENU_OPEN: 'MENU_OPEN',
+} as const
+
+export type IconKind = typeof ICON_KIND[keyof typeof ICON_KIND]
